2008-04-10  Hidehisa Akiyama  <akky@users.sourceforge.jp>
	    * qt4/qt4.pro:
	    * classic/Makefile.am:
	    * classic/logplayer.cpp:
	    * configure.ac:
	    * qt4/Makefile.am:
	    * qt4/team_graphic_painter.cpp:
	    * rcsslogplayer/rcsslogplayer.pro:
	    * rcsslogplayer/util.cpp:
	    - fixed build problem on Mac OS X, but cannot be build by
	    Makefile generated by configure. instead of them, use qmake.



git-svn-id: file:///home/akiyama/sserver/svn/rcss/trunk/rcsslogplayer@267 1639d2a8-b03b-4162-8b0b-03f2cde17881

+++ qt4/team_graphic_painter.cpp
@@ -170,7 +170,7 @@ TeamGraphicPainter::copyTeamGraphicXpmTile( QPixmap & dst_pixmap,
     xpm[0] = new char[64];
     std::snprintf( xpm[0], 64, "%d %d %d %d",
                    tile.width(), tile.height(),
-                   tile.colors().size(), tile.cpp() );
+                   static_cast< int >( tile.colors().size() ), tile.cpp() );
 
     // colors
     std::size_t idx = 1;
